+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - have a progress bar for long-running commands when possible
 - use [gspread](https://docs.gspread.org/) to handle the low-level GoogleAPI
 - use [gspread-dataframe](https://gspread-dataframe.readthedocs.io/) to handle converting a worksheet into a dataframe
+- timeout of 60s for PretalxAPI as it is really slow which caused a lot of timeout errors
 
 ## Version 0.1.1 (2023-01-16)